remove fullscreen - it's not reliable
It seems quite difficult and annoying to get "true fullscreen" working
properly on the high-DPI macbooks. As of now, it's much more reliable to
use the green "expand window" button than start fullscreen
programatically. Therefore, we remove the feature. The branch
`true_fullscreen` has an in-progress implementation of reliable
fullscreen, but it's still not good enough.